server/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
cloup==2.1.2
decorator==5.1.1
Django==4.2.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
glcontext==2.5.0
gunicorn==21.2.0
idna==3.6
isosurfaces==0.1.0
manim==0.18.0
ManimPango==0.5.0
mapbox-earcut==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
moderngl==5.10.0
moderngl-window==2.4.4
multipledispatch==1.0.0
networkx==3.2.1
numpy==1.26.4
packaging==23.2
Pillow==9.5.0
pycairo==1.26.0
pydub==0.25.1
pyglet==2.0.12
Pygments==2.17.2
pyrr==0.10.3
pytz==2024.1
requests==2.31.0
rich==13.7.1
scipy==1.12.0
screeninfo==0.8.1
shapely==2.0.3
skia-pathops==0.7.4
sqlparse==0.4.4
srt==3.5.3
svgelements==1.9.6
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
watchdog==3.0.0
